Split browser frame rendering modules
