+++ src/app_core/native_bridge/tests/mod.rs
@@ -94,8 +94,10 @@ fn runtime_exit_returns_structured_shutdown_timing_once() {
 
 #[test]
 fn runtime_exit_detaches_active_controller_shutdown_work() {
-    const BLOCKING_FILE_OP: Duration = Duration::from_secs(2);
-    const DETACHED_EXIT_BUDGET: Duration = Duration::from_millis(1_000);
+    /// Long enough for the test file operation to prove shutdown does not wait on it.
+    const BLOCKING_FILE_OP: Duration = Duration::from_secs(10);
+    /// Upper bound for the detached runtime-exit handoff.
+    const DETACHED_EXIT_BUDGET: Duration = Duration::from_secs(5);
 
     let base = tempdir().expect("create temp config dir");
     let _base_guard = crate::app_dirs::ConfigBaseGuard::set(base.path().to_path_buf());
